+++ test/childClient.js
@@ -35,8 +35,8 @@ describe('#childClient', () => {
           globalTags: ['awesomeness:over9000', 'tag1:xxx', 'bar', ':baz']
         });
 
-        assert.strictEqual(child.prefix, 'preff.prefix');
-        assert.strictEqual(child.suffix, 'suffix.suff');
+        assert.strictEqual(child.prefix, 'preff.prefix.');
+        assert.strictEqual(child.suffix, '.suffix.suff');
         assert.strictEqual(statsd, global.statsd);
         assert.deepEqual(child.globalTags, ['gtag', 'awesomeness:over9000', 'tag1:xxx', 'bar', ':baz']);
       });